chore: Move to poetry, clean up some files, add CI workflow
We are duplicating a lot of things within the project setup. This
is the first PR to move us into a more standard setup. We don't
need setup.py or requirements.txt. This can all live within the
project.toml. The PR tries to limit changing everything at once.

Main updates are:
1. Move requirements out of requirements.txt to pyproject. This allows
dependencies to actually be installed during pip installs.
2. Cleans up some formatting, unused imports, etc.
3. Removes unneeded dependencies.
4. Fixes a few failing tests that were not working prior.
5. Adds a "pr_build" GHA, so tests are constantly running.