chore: Update docs builds within build-deploy-all GHA
We no longer use github pages for docs and instead have them
within the Kitops repo/site. We remove the upload to github pages
from the build-deploy-all GHA and instead call the sync-docs-to-kitop
repo.

The sync-docs-to-kitops GHA no longer triggers on upload to main
and instead is a `workflow_call`, which allows another GHA to call
it. In addition, we added `workflow_dispatch` to allow maintainers
to deploy docs as needed. The worflow itself remains the same.